requirements.txt
django
django-simple-captcha
celery
redis
lxml
